ref:move configuration providers to right places
