+++ Src/Campus.Services.Interfaces/Interfaces/Configuration/IConfigurationProviderFactory.cs
@@ -1,4 +1,4 @@
-namespace Campus.Domain.Core.Interfaces
+namespace Campus.Services.Interfaces.Interfaces.Configuration
 {
     public interface IConfigurationProviderFactory
     {
