+++ Src/Campus.Services/Infrastructure/ProfileService.cs
@@ -1,6 +1,5 @@
 using System;
 using System.Collections.Generic;
-using System.Threading;
 using System.Threading.Tasks;
 using Campus.Domain.Core.Models;
 using Campus.Infrastructure.Data.EntityFrameworkCore.Context;
@@ -9,7 +8,7 @@ using Campus.Services.Interfaces.Interfaces;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.EntityFrameworkCore;
 
-namespace Campus.Services.Infrastructure
+namespace Campus.Services.Implementation.Infrastructure
 {
     public class ProfileService : IProfileService
     {
